+++ testdata/mod3/mydir/dir.go
@@ -1,0 +1,9 @@
+package dir
+
+import (
+	"github.com/ethereum/go-ethereum/log"
+)
+
+func Dir() {
+	_ = log.StdoutHandler
+}