+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Most recent version is listed first.
   In repositories that have nested Go modules; ignore all the nested modules and only   
   work/analyze the module that is passed in as an argument to `ote`
 - Have all static analysis passes succeed: https://github.com/komuw/ote/pull/20
+- dont analyze Go files inside vendor/ directory: https://github.com/komuw/ote/pull/21
 
 
 ## v0.0.3