+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Most recent version is listed first.
 - Have all static analysis passes succeed: https://github.com/komuw/ote/pull/20
 - dont analyze Go files inside vendor/ directory: https://github.com/komuw/ote/pull/21
 - render unformatted `//test` comment correctly: https://github.com/komuw/ote/pull/24
+- perf improvement, do not generate a new string: https://github.com/komuw/ote/pull/25
+  This is an improvement of https://github.com/komuw/ote/pull/24
 
 
 ## v0.0.3