+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Most recent version is listed first.
 - perf improvement, do not generate a new string: https://github.com/komuw/ote/pull/25
   This is an improvement of https://github.com/komuw/ote/pull/24
 - call `fetchModule` only for the import paths that are not shared between test files and non-test files: https://github.com/komuw/ote/pull/26
--
+- add more tests: https://github.com/komuw/ote/pull/29
 
 
 ## v0.0.3